Backbone.Brace -- a second pass at declarative event routing.
Based on feedback from bgrins and jashkenas I've improved the approach with bindings to be more flexible and configurable, while still providing a conventional approach if desired.

The idea here is that Brace provides a default routing scheme that can be overridden with a router function and or a routing table. Brace maintains a registry of these routes so that the bindings can be released easily. The design of this is now orthogonal to Backbone.View. Brace could be mixed in directly to any object by the user such as a controller or custom view just as easily as it can be used inside Backbone.View.

The name choice is probably a little too cute, but I thought I'd throw it out there.

Let me know if you think this is an improvement? Thanks for your feedback.