Added catchup support for late binding events.
Now when events are bound to, if they enable catchup, the bound callback will
be triggered for the # times the event has occured in the past. This will allow
for event callbacks suck as counters, to be late bound, but still receive a
callback for each event triggering that they missed.