Backbone View to use jQuery object for this.el, instead of DOM object Edit Change all references of "el" in the view from a DOM element to a jQuery/Zepto element. This change can potentially break code and as such could require code migrations
Use case:
Because we're requiring jQuery/Zepto by default (which is a DOM abstraction lib), all of our DOM interactions should be delegated to jQuery to normalise.
All of the code creating that extends view will use $(this.el) which is inifficient, it is much better to do this in the view base layer, so that all extended code can use this.el as the jQuery instantiated object.