beforeSave/afterSave around Backbone.Model#save
beforeSave receives a reference to the model and the options object. The
handler can adjust the options hash before call the Backbone.sync

afterSave receives a success flag (true or false), which indicates if the
response was successful or not, and a reference to the model. The event
is triggered before Model#set.