view.el factory/selector function
"You may assign el directly if the view is being created for an element that already exists in the DOM. Use either a reference to a real DOM element, a function that returns a DOM element, or a css selector string."

A view.el function that returns a DOM element is useful for bundling templating code or advanced selectors within the View.

Tests included. I can go deeper and update the documentation if you think you'll pick up this pull request.