added support for isNew option when instantiating a model
when a model is instantiated, an option isNew can be included with a
value of true.  If isNew is true then the model will be considered to
be new regardless of whether id is null or not.  This is helpful for
tables that do not use an auto-increment as the primary key