Introduced Child-Views to Backbone.View
The method attachChildViews can be used to attach views to a parent-view. When the parent-view gets disposed by calling myView.destroy() the child-views will get destroyed, too. 

The outitialize-method is the opposite of initialize and can be used to unbind views from events when the view is destroyed. (It's like a onDestroy-Method, but I like the name "outitialize" better for symmetrical reasons :-)

Any opinions on this commit? It's my first contribution on github. Please point me in the right direction in case I made any mistakes or broke any rules.

Thanks :-)

Tobias