Allow HTTP verbs to be specified for sync
While the CRUD->REST mapping can be useful conceptually, it's not entirely accurate.  Because POST is not idempotent, but PUT is, there are times a create or an update could be either.

This change allows HTTP verbs to be directly passed to sync, ignoring the mapping.

A possible alternative (which could also be used in conjunction with this change), would be to add 'get', 'put', 'post', and 'delete' to the methodMap.