Document how view constructor handles options
Closes #1749

Backbone.View calls the _configure function which uses _.extend to extend
the properties of an options property declared on the view (if it exists).
This provides a mechanism for defining default view options that can be
overridden on a per-instance basis (in a similar fashion to how the el,
id and className options are handled).  This functionality was previously
undocmented.