Import _ and $ from the true global object instead of root (fixes #1789)
When Backbone is loaded as a module in the browser, e.g. by require.js, the value of `this` at the top level may be different from `window` (a common choice is `this === exports`).

When `_` or `$` are defined globally, the right place to look for them is on the true global object, and the most reliable way to retrieve that object (even in ES5 strict mode) is by evaluating `Function('return this')()`.