Signed-off-by: Adriaan Labuschagne <adriaanlabuschagne@gmail.com>
All events model.set calls that cause a change to be added to the changes array are now added up
in the this._pending variable. This allows the original model.set that started a chain of sets (through
triggering events) to fire a one 'change' event for each set that caused a change to be added to the 'changes' array.
This fixes issue #2034