Follow OWASP guidelines for bootstrapping data in documentation.
Current documentation suggests writing initial collection data to a
SCRIPT element. The documentation suggests further escaping to prevent
injections from untrusted data. The best practice for this is to write
the JSON into a hidden div, read the JSON out as text, and use
JSON.parse to read it into the context.