Removed hard coded Backbone.history dependency in Router
Backbone assumes that History is a singleton which is not the case in
non-browser environments like node.js where multiple requests have to
be handled simultaneously. This commit is not an attempt to rethink how
Backbone handles Controller/Dispatcher logic across different
environments but a merely an improvement that allows using Router and
History on the server side without breaking the existing api.