Transition from hash-based route: check it's really a route (and not an anchor)
On `http://mysuperapp.com` suppose we have a `<a href="#top">Go to top</a>`. After having clicked on it, location is now `http://mysuperapp.com/#top`.

When refreshing the page (F5), Backbone history converts it to `http://mysuperapp.com/top` which is wrong because `'top'` is not a route.

Just check the fragment is **really** a route prevents from this and allows usage of anchors in pushState-enabled browsers.