coddd documentation to delegateEvents for cases of rendering child views
Calling jQuery’s `html()` method in a parent view internally calls
`empty()`, which clears an element’s inner nodes, losing any bound
events on child views. This PR adds documentation to `delegateEvents`
to describe this use case, including how to maintain events on a child
view when the parent view is re-rendered, and includes an example.