Align few places in code to match with possible overriding of Backbone.sync
Overriding Backbone.sync is a great way to switch a persistence layer. It just lacks two moments that should be overridable too:

* Result of `model.save` if model was not validated
* Result of `model.destory` if model was new

For example if my implementation of Backbone.sync returns promises, then current implementation forces me to check the result before calling '.then()'. With the proposed changes, we will be able to unify `sync` behavior with the rest of the Backbone.

@jashkenas ?