adding feedback to update event of collection
collection's update event receives 2 additional attributes, in case of removal of existing models, previosModels is an array containing the removed models which where references to the collection before .set was called. The second attribute, addedModels contains an array of models which were added to the collection calling the .set method.

Signed-off-by: elgubenis <elgubenis@gmail.com>