Clarify Router.navigate docs
Issue 3183. Clarify the use `{ replace: true }` with Router.navigate - that is if you do not want the current url in history that you should use this option when navigating away from that route. The current documentation led me to believe that it was for navigating to a route that I did not want in history.